logging-service/requirements.txt
grpcio
grpcio-tools
grpcio-reflection
protobuf
hazelcast-python-client
flask
requests